% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{add_bold}
\alias{add_bold}
\title{Make a string bold in a certain format}
\usage{
add_bold(str, output_type)
}
\arguments{
\item{str}{string or strings (character) to make bold}

\item{output_type}{output type (see knitr::kable's format)}
}
\description{
For example, in latex it would take "2" and output "{\\bf 2}"; in html
it would output "<b>2</b>".
}
