% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-to-simulation.R
\name{add_dref_to_list}
\alias{add_dref_to_list}
\title{Internal function to add DrawsRef to a list of DrawsRef objects}
\usage{
add_dref_to_list(dref, dref_list, sim_dir)
}
\arguments{
\item{dref}{DrawsRef to add}

\item{dref_list}{list of DrawsRef objects}

\item{sim_dir}{sim@dir}
}
\description{
Makes sure that DrawsRef with this same index is not already in list.
Although not checked, it is assumed that this is only called on a list
of DrawsRef objects all coming from same model.
}
\keyword{internal}

