% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-to-simulation.R
\name{add_eref_to_list}
\alias{add_eref_to_list}
\title{Internal function to add EvalsRef to a list of EvalsRef objects}
\usage{
add_eref_to_list(eref, eref_list, sim_dir)
}
\arguments{
\item{eref}{EvalsRef to add}

\item{eref_list}{list of EvalsRef objects}

\item{sim_dir}{sim@dir}
}
\description{
Makes sure that EvalsRef with this same index and method is not already in
list. Although not checked, it is assumed that this is only called on a list
of EvalsRef objects all coming from same model.
}
\keyword{internal}
