% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-to-simulation.R
\name{add_oref_to_list}
\alias{add_oref_to_list}
\title{Internal function to add OutputRef to a list of OutputRef objects}
\usage{
add_oref_to_list(oref, oref_list, sim_dir)
}
\arguments{
\item{oref}{OutputRef to add}

\item{oref_list}{list of OutputRef objects}

\item{sim_dir}{sim@dir}
}
\description{
Makes sure that OutputRef with this same index and method is not already in
list. Although not checked, it is assumed that this is only called on a list
of OutputRef objects all coming from same model.
}
\keyword{internal}
