% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregator-class.R
\name{aggregate_evals}
\alias{aggregate_evals}
\title{Apply aggregator to a list of Evals objects}
\usage{
aggregate_evals(evals_list, aggregator)
}
\arguments{
\item{evals_list}{a list of Evals objects}

\item{aggregator}{object of class Aggregator}
}
\description{
Returns a num_models by num_methods matrix
}

