% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregator-class.R
\docType{class}
\name{Aggregator-class}
\alias{Aggregator-class}
\title{An S4 class for aggregating evaluated metrics}
\description{
An object of class \code{Aggregator} consists of a label and a function
\code{aggregate} that has a single argument \code{ev} that is a list of
length \code{nsim}.  This list consists of the evaluated values of a single
metric on a single method for a single model.
}
\section{Slots}{

\describe{
\item{\code{label}}{a human readable label that will be a prefix to the Eval's label}

\item{\code{aggregate}}{a function with argument \code{ev} that is a list of
length \code{nsim} and returns a scalar.}
}}

