% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening.R
\name{all_picker}
\alias{all_picker}
\title{Picks all rows every time}
\usage{
all_picker(state)
}
\arguments{
\item{state}{as defined by \code{make_susceptible} function}
}
\value{
vector of TRUE/FALSE picking rows of the state for testing
}
\description{
Picks all rows every time
}
