% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evals-class.R
\name{as.data.frame.Evals}
\alias{as.data.frame.Evals}
\title{Convert an Evals to a data.frame}
\usage{
\method{as.data.frame}{Evals}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{object of class \code{\linkS4class{Evals}}}

\item{row.names}{not used}

\item{optional}{not used}

\item{...}{not used}
}
\description{
This is equivalent to calling \code{as(x, "data.frame")}
}
