% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\name{as.data.frame.listofModels}
\alias{as.data.frame.listofModels}
\title{Convert a List of Models to a data.frame}
\usage{
\method{as.data.frame}{listofModels}(x, row.names = NULL, optional = FALSE,
  ...)
}
\arguments{
\item{x}{list}

\item{row.names}{not used}

\item{optional}{not used}

\item{...}{not used}
}
\description{
When \code{\link{load}} generates a list of Models, it assigns this
to be of (S3) class listofModels, inherited from list, so that this function
will be invoked instead of as.data.frame.list, which is defined in base.
}

