% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\name{as.data.frame.Model}
\alias{as.data.frame.Model}
\title{Convert a Model to a data.frame}
\usage{
\method{as.data.frame}{Model}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{object of class \code{\linkS4class{Model}}}

\item{row.names}{not used}

\item{optional}{not used}

\item{...}{not used}
}
\description{
Ignores any params that are not length 1 and numeric or character. This is
equivalent to calling \code{as(x, "data.frame")}
}
