% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\name{$,Model-method}
\alias{$,Model-method}
\title{Get element of \code{\linkS4class{Model}}'s \code{params} list}
\usage{
\S4method{$}{Model}(x, name)
}
\arguments{
\item{x}{object of class \code{\linkS4class{Model}}}

\item{name}{name of an element appearing in \code{x@params}}
}
\description{
Get element of \code{\linkS4class{Model}}'s \code{params} list
}
