% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\docType{data}
\name{categorical_classes}
\alias{categorical_classes}
\title{Classes used to indicate categories}
\format{
An object of class \code{character} of length 3.
}
\usage{
categorical_classes
}
\description{
Classes used to indicate categories
}
\keyword{datasets}
