% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{catsim}
\alias{catsim}
\title{Concatenate and print for the simulator}
\usage{
catsim(...)
}
\arguments{
\item{...}{arguments to be passed to \code{\link{cat}}}
}
\description{
For internal use.  This calls \code{\link{cat}} only when
\code{getOption("simulator.verbose")}.
}
