% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component-class.R
\docType{class}
\name{Component-class}
\alias{Component-class}
\title{An S4 class representing a component of the simulator.}
\description{
This is a virtual class.
}
\section{Slots}{

\describe{
\item{\code{name}}{a short name identifier.  Must be alphanumeric.}

\item{\code{label}}{a longer, human readable label that can have other characters
such as spaces, hyphens, etc.}
}}

