% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{count_states}
\alias{count_states}
\title{Summarize state}
\usage{
count_states(state)
}
\arguments{
\item{state, }{as describe by \code{make_susceptible_state} function}
}
\value{
count of S, E, I, and R state by time, location, and quarantine
state
}
\description{
Summarize state
}
