% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create}
\alias{create}
\title{Create template for a new set of simulations}
\usage{
create(dir = "./my_sims")
}
\arguments{
\item{dir}{where to create the skeleton of a new set of simulations}
}
\description{
This function is the fastest way to get started.  Creates the skeleton of a
simulation.
}
\examples{
\dontrun{
 create("./examples")
 }
}

