% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counters.R
\name{ctr}
\alias{ctr}
\title{Counters that can be shared among objects.}
\description{
Counters that can be shared among objects.

Counters that can be shared among objects.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{increment}}{return the next value from the counter}

\item{\code{current}}{return the current counter value}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ctr$new()}}
\item \href{#method-incr}{\code{ctr$incr()}}
\item \href{#method-clone}{\code{ctr$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a counter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctr$new(i = 0L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{start counting at i}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-incr"></a>}}
\if{latex}{\out{\hypertarget{method-incr}{}}}
\subsection{Method \code{incr()}}{
Increment the counter explicitly
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctr$incr()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctr$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
