% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{default_cluster_name}
\alias{default_cluster_name}
\title{Package-specific location for holding the cluster reference}
\usage{
default_cluster_name()
}
\value{
name of variable holding a cluster reference
}
\description{
Package-specific location for holding the cluster reference
}
