% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening.R
\name{default_test}
\alias{default_test}
\title{Something like a PCR test}
\usage{
default_test(state, picks)
}
\arguments{
\item{state}{as defined by \code{make_susceptible} function}

\item{picks}{rows to include in testing, 0 = skip, > 0 = test}
}
\value{
state with modified \code{test_outcome} and \code{last_test} states
}
\description{
Something like a PCR test
}
