% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{describe}
\alias{describe}
\title{Describe the contents of a simulator directory}
\usage{
describe(dir = ".")
}
\arguments{
\item{dir}{name of the directory where directory named "files" exists}
}
\description{
Describe the contents of a simulator directory
}
