% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counters.R
\docType{data}
\name{.simulator_counters}
\alias{.simulator_counters}
\title{Location where global counters are stored}
\format{
An object of class \code{environment} of length 1.
}
\usage{
.simulator_counters
}
\description{
Each named counter is independent of all other counters with different names
}
\keyword{datasets}
