% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-env.R
\name{env_list_as_tibble_row}
\alias{env_list_as_tibble_row}
\title{Turn a list of enivronments into a tibble row}
\usage{
env_list_as_tibble_row(p)
}
\arguments{
\item{p}{list of environments}
}
\value{
tibble row
}
\description{
Resulting tibble includes all variables from all environments in a flat
structure
}
