% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-env.R
\name{env_list_list_as_tibble}
\alias{env_list_list_as_tibble}
\title{Turn a list of environment lists into a tibble}
\usage{
env_list_list_as_tibble(l)
}
\arguments{
\item{l}{list of lists of environments}
}
\value{
a tibble
}
\description{
Turn a list of environment lists into a tibble
}
