% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate_single}
\alias{evaluate_single}
\title{Run one or more metrics on outputs.}
\usage{
evaluate_single(metrics, model, output, draws = NULL)
}
\arguments{
\item{metrics}{a list of \code{\linkS4class{Metric}} objects}

\item{model}{a \code{\linkS4class{Model}} object}

\item{output}{a \code{\linkS4class{Output}} object}

\item{draws}{(optional) a \code{\linkS4class{Draws}} object or NULL}
}
\description{
This is an internal function.  Users should call the wrapper function
\code{\link{evaluate}}. Here "single" refers to a single output (and
thus a single method, though not necessarily a single index).
The metrics provided are run and saved together in a file.
}

