% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{expand_parameter_grid}
\alias{expand_parameter_grid}
\title{Expand parameter grid}
\usage{
expand_parameter_grid(x)
}
\arguments{
\item{x}{a list, optionally with nested lists, all named elements,
and leaf elements must be vectors of 'numeric', 'integer', or
'character' class.  They are expanded to all combinations
in blocks (nested lists create blocking).}
}
\value{
list of tibbles, with nesting in lists
}
\description{
Inputs are a list of (optionally nested lists of) vectors and these are
expanded within-list using expand.grid and transformed to tibbles for
consistency.  Nesting remains to avoid creating the full (possible large)
tibble that represents the full parameter grid.
}
