% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{expand_parameters}
\alias{expand_parameters}
\title{Expand lists of parameters, combine and insert metadata}
\usage{
expand_parameters(
  ...,
  recipes,
  n_batches = 50,
  timestamp = timestamp(),
  revision = ""
)
}
\arguments{
\item{...}{each dot-arg is a list of (optionally nested) lists
of (ultimately) numeric/integer/character arguments or
tibbles.}

\item{recipes}{used to create derived parameters from the primary
expanded parameters}

\item{n_batches}{number of batches to break the work into (only
generates a grouping factor with labels).}
}
\value{
tibble of parameter values
}
\description{
Each of the dot-args is exapnded individuall, they are combined
using bind_rows and so much have matching parameters, and then
simulation id's, replicate id's, job tags, and job batch labels
are added
}
