% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{expand_tibbles}
\alias{expand_tibbles}
\title{Expand a flat list of tibbles}
\usage{
expand_tibbles(x)
}
\arguments{
\item{x}{a flat list of tibbles}
}
\value{
a single tibble with (likely) nested tibbles
}
\description{
The implied expansion is for each subsequent tibble to be inserted
as a list column into the first tibble.  This is the equivalent to
expand.grid on a (flat) list of tibbles).
}
