% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{expand_tibbles_nested}
\alias{expand_tibbles_nested}
\title{Flatten a list of (possibly nested list of) tibbles to a single tibble}
\usage{
expand_tibbles_nested(x)
}
\arguments{
\item{a}{list of (possibly nested lists of ) tibbles}
}
\value{
a single tibble with nested columns
}
\description{
Nesting is preserved as nested list columns of tibbles
}
