% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{find_betweens}
\alias{find_betweens}
\title{Finding numbers between
find an equal number of equidistant points between each number in x
to get a total n number of numbers.
any extra points will be made where the interval (distance between two #s in x)
is largest.}
\usage{
find_betweens(x, n)
}
\arguments{
\item{x}{a vector of numbers}

\item{n}{number of total points}
}
\value{
a vector of length n
}
\description{
Finding numbers between
find an equal number of equidistant points between each number in x
to get a total n number of numbers.
any extra points will be made where the interval (distance between two #s in x)
is largest.
}
