% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{flatten_tibble}
\alias{flatten_tibble}
\title{Unnest all columns within a nested tibble recursively}
\usage{
flatten_tibble(x)
}
\arguments{
\item{x}{a tibble}
}
\value{
a tibble with no nested columns left
}
\description{
Unnest all columns within a nested tibble recursively
}
