% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{flatten_tibble_list}
\alias{flatten_tibble_list}
\title{Flatten nested list of tibble to non-nested list of tibbles}
\usage{
flatten_tibble_list(x, l = list())
}
\arguments{
\item{x}{list of nested (sometimes lists of) tibbles flatten}

\item{l}{propagate return value in recursion}
}
\value{
list of tibbles
}
\description{
Flatten nested list of tibble to non-nested list of tibbles
}
