% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{generate_jobs}
\alias{generate_jobs}
\title{Turn inputs into a set of job descriptors}
\usage{
generate_jobs(
  parameters = list(),
  initialization = recipe(),
  step = recipe(),
  writer = list(factory = make_writer, setup = recipe(), summaries = recipe())
)
}
\arguments{
\item{parameters}{parameters defining the simulation}

\item{initialization}{recipe for initializing simulation}

\item{step}{recipe for each simulation step}

\item{writer}{list with elements anmed 'factory', 'setup', 'summaries'}
}
\value{
a list ...
}
\description{
Turn inputs into a set of job descriptors
}
