% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{get_cluster}
\alias{get_cluster}
\title{Automatically retrieve \emph{some} cluster that is running}
\usage{
get_cluster(
  cl = NULL,
  variable = default_cluster_name(),
  env = .GlobalEnv,
  ...
)
}
\arguments{
\item{cl}{object maybe holding a cluster}

\item{variable}{name of variable holding cluster}

\item{env}{where the object holding the cluster (reference) will live}
}
\value{
a running cluster from parallel package
}
\description{
Automatically retrieve \emph{some} cluster that is running
}
