% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counters.R
\name{get_counter}
\alias{get_counter}
\title{Retrieve a global counter}
\usage{
get_counter(name = "default_counter")
}
\arguments{
\item{name}{name of the counter}
}
\value{
a counter
}
\description{
Retrieve a global counter
}
