% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{get_files_not_in_simulations}
\alias{get_files_not_in_simulations}
\title{Find files in simulator directory not referred to by any simulations}
\usage{
get_files_not_in_simulations(dir, out_loc = "out")
}
\arguments{
\item{dir}{name of the directory where directory named "files" exists}

\item{out_loc}{a length-1 character vector that gives location
(relative to model's path) that method outputs are stored.This can be
useful for staying organized when multiple simulations are based on
the same Model and Draws objects.  Usually this is just "out"}
}
\description{
Once one has completed all simulation studies, this function can be called
to identify any files that may have been created along the way that are no
longer being used in any simulations.  It would then be safe to delete these
files.
}
