% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-from-simulation.R
\name{get_model_indices}
\alias{get_model_indices}
\title{Returns indices of a specified subset of sim@model_refs}
\usage{
get_model_indices(sim, subset)
}
\arguments{
\item{sim}{a simulation object}

\item{subset}{a vector indicating which models should be returned.}
}
\description{
See \code{\link{model}} for information about the various formats of subset.
}
