% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_relative_path}
\alias{get_relative_path}
\title{Get relative path}
\usage{
get_relative_path(base_path, path)
}
\arguments{
\item{base_path}{the base path}

\item{path}{a specific path}
}
\description{
Given a base path and a specific path, returns a string str such that
file.path(base_path, str) is the same location as path.
}
