% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{get_state_summaries}
\alias{get_state_summaries}
\title{Get summaries}
\usage{
get_state_summaries(output_dir)
}
\arguments{
\item{output_dir}{every summary under this root will be indexed}
}
\value{
a data frame listing all summaries with their set and replicate
index
}
\description{
Returns the summarized states for all sets/replicates
}
