% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timings.R
\name{get_timings}
\alias{get_timings}
\title{Get timings for all jobs recursively}
\usage{
get_timings(output_dir)
}
\arguments{
\item{root}{dir to search from}
}
\value{
tibble with all timing data
}
\description{
Get timings for all jobs recursively
}
