% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eval.R
\name{ggplot_eval}
\alias{ggplot_eval}
\title{Make a boxplot of a metric for each method using ggplot2}
\usage{
ggplot_eval(
  evals_df,
  metric_name,
  method_name,
  method_label,
  main,
  facet_mains,
  ylab,
  ylim,
  nrow,
  ncol
)
}
\description{
Make a boxplot of a metric for each method using ggplot2
}
\keyword{internal}
