% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection.R
\name{infection}
\alias{infection}
\title{Compartment transfers of SEIR model}
\usage{
infection(state, risk)
}
\arguments{
\item{state}{as defined by \code{make_susceptible} function}

\item{risk}{risk table from \code{risk} function}
}
\value{
state, with modified SEIR memership, no other column changes
}
\description{
Compartment transfers of SEIR model
}
