% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\docType{methods}
\name{load,list-method}
\alias{load,list-method}
\title{Load a list of reference objects}
\usage{
\S4method{load}{list}(file)
}
\arguments{
\item{file}{list of objects to load}
}
\description{
Load a list of reference objects
}

