% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\docType{methods}
\name{load,ModelRef-method}
\alias{load,ModelRef-method}
\title{Load a ModelRef}
\usage{
\S4method{load}{ModelRef}(file)
}
\arguments{
\item{file}{object to load}
}
\description{
Load a ModelRef
}

