% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\docType{methods}
\name{load,OutputRef-method}
\alias{load,OutputRef-method}
\title{Load an OutputRef}
\usage{
\S4method{load}{OutputRef}(file)
}
\arguments{
\item{file}{object to load}
}
\description{
Load an OutputRef
}

