% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{make_my_example_model}
\alias{make_my_example_model}
\title{Make My Example Model}
\usage{
make_my_example_model(n)
}
\arguments{
\item{n}{number of i.i.d. draws}
}
\description{
This function is used in the examples.  It returns a
\code{\linkS4class{Model}} object.  In particular, it represents
\code{n} i.i.d. draws from a normal with mean 2 and variance 1.
}
\seealso{
\code{\link{my_example_method}} \code{\link{my_example_loss}}
}

