% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-helpers.R
\name{make_population}
\alias{make_population}
\title{Create the initial state for a simulation.}
\usage{
make_population(time, group, ..., steps = recipe(), .env, .counter)
}
\arguments{
\item{time}{base time to start pop from}

\item{group}{group name for this pop}

\item{...}{data available to parameterize recipe}

\item{steps}{recipe to construct population columns}

\item{.env}{enviornment use for population}

\item{.counter}{counter used as source for id's}
}
\value{
tibble::tibble holding the simulation state for N individuals
}
\description{
This is also an informal definition of the format of data stored
for the simulation.  We avoid ad-hoc modifications of the columns in the
code.  In the future that should make it possible to make this more of an
insulated object.
}
\details{
Each row holds the state for one individual, although having it hold the
state for one group is reasonable by turning the 0/1 convention into a count
convention.  Marking the functions that would need to change elsewhere with
FIXME as a go.
}
