% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{memory_as_string}
\alias{memory_as_string}
\title{Write memory in human readable way}
\usage{
memory_as_string(memory_in_bytes)
}
\arguments{
\item{memory_in_bytes}{the amount of memory in Bytes.}
}
\description{
Write memory in human readable way
}

