% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric-class.R
\docType{class}
\name{Metric-class}
\alias{Metric-class}
\title{An S4 class representing an evaluation metric to be used by simulator.}
\description{
An object of class \code{Metric} consists of a name, label, and a function
\code{metric} that takes arguments \code{model} (of class
\code{\linkS4class{Model}}) and \code{out} (of class \code{\linkS4class{Output}}), which
is the output of a method.
}
\details{
This class inherits from the \code{\linkS4class{Component}} class.
}
\section{Slots}{

\describe{
\item{\code{name}}{a short name identifier.  Must be alphanumeric.}

\item{\code{label}}{a longer, human readable label that can have other characters
such as spaces, hyphens, etc.}

\item{\code{metric}}{a function with arguments "model" and "out" (and optionally
"draw")}
}}

