% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{mix_generator}
\alias{mix_generator}
\title{Generate a mix of states}
\usage{
mix_generator(n, mix, labels, .state = TRUE)
}
\arguments{
\item{n}{number of units in the vector}

\item{mix}{named simplex of proportions of each state}

\item{labels}{additional (or primary) labels, optional}

\item{.state}{dev hook, return a character vector if FALSE, otherwise
a 'state' class}
}
\value{
state vector sampled from mix, format depends on .state
}
\description{
Generate a mix of states
}
