% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-class.R
\name{model_names}
\alias{model_names}
\title{Get model names in a Simulation}
\usage{
model_names(sim)
}
\arguments{
\item{sim}{object of class \code{\linkS4class{Simulation}}}
}
\description{
Get model names in a Simulation
}
