% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference-classes.R
\docType{class}
\name{ModelRef-class}
\alias{ModelRef-class}
\title{An S4 class representing a reference to an object of class Model.}
\description{
This identifies the necessary information to locate a saved object
of class \code{\linkS4class{Model}}.
}
\section{Slots}{

\describe{
\item{\code{dir}}{directory where the directory "files" is that contains the
referenced \code{\linkS4class{Model}} object}

\item{\code{name}}{a short name identifier.}

\item{\code{label}}{a longer, human readable label that can have other characters}

\item{\code{simulator.files}}{simulator functions will use
\code{getOption("simulator.files")} if simulator.files not provided.}
}}

