% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\name{models_as_data.frame}
\alias{models_as_data.frame}
\title{Convert a list of Model objects into a data.frame}
\usage{
models_as_data.frame(m)
}
\arguments{
\item{m}{model object}
}
\description{
Ignores any params that are not length 1 and numeric or character
}

