% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\docType{data}
\name{my_example_loss}
\alias{my_example_loss}
\title{My Example Loss}
\format{An object of class \code{Metric} of length 1.}
\usage{
my_example_loss
}
\description{
This \code{\linkS4class{Metric}} object is used in the examples.  It is
squared error loss.
}
\seealso{
\code{\link{make_my_example_model}} \code{\link{my_example_loss}}
}
\keyword{datasets}

