% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\docType{data}
\name{my_example_method}
\alias{my_example_method}
\title{My Example Method}
\format{
An object of class \code{Method} of length 1.
}
\usage{
my_example_method
}
\description{
This \code{\linkS4class{Method}} object is used in the examples.  It is
the sample mean of the data.
}
\seealso{
\code{\link{make_my_example_model}} \code{\link{my_example_loss}}
}
\keyword{datasets}
