% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregator-class.R
\name{new_aggregator}
\alias{new_aggregator}
\title{Create an Aggregator object}
\usage{
new_aggregator(label, aggregate)
}
\arguments{
\item{label}{a human readable label}

\item{aggregate}{a function with argument \code{ev} that is a list of
length \code{nsim} with each element itself being a named list. Each
element of this list corresponds to a metric that has been computed.
In particular, given an \code{\linkS4class{Evals}} object \code{o},
\code{aggregate} takes as input \code{o@evals[[method_name]]} (which
is a list of the kind just described).
The function aggregate should return a scalar.}
}
\description{
Creates a new \code{\linkS4class{Aggregator}} object.
}

