% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-class.R
\name{new_method}
\alias{new_method}
\title{Create a Method object}
\usage{
new_method(name, label, method, settings = list())
}
\arguments{
\item{name}{a short name identifier.  Must be alphanumeric.}

\item{label}{a longer, human readable label that can have other characters
such as spaces, hyphens, etc.}

\item{method}{a function that has arguments "model", "draw"
and (optionally) names matching elements within \code{names(settings)}}

\item{settings}{(optional) a list of "settings" for the method (e.g., tuning
parameters or related information that might distinguish two otherwise
identical methods).}
}
\description{
Creates a new \code{\linkS4class{Method}} object.
}
