% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-extension-class.R
\name{new_method_extension}
\alias{new_method_extension}
\title{Create an object that can be used to make an extended version of a method}
\usage{
new_method_extension(name, label, method_extension)
}
\arguments{
\item{name}{a short name identifier.  Must be alphanumeric.}

\item{label}{a longer, human readable label that can have other characters
such as spaces, hyphens, etc.}

\item{method_extension}{a function with arguments "model", "draw", "out", and
"base_method".  This will become the function \code{extended_method}
in the \code{ExtendedMethod} object that is created.}
}
\description{
Creates an object of class \code{MethodExtension}, which when added to a
\code{Method} creates an \code{\linkS4class{ExtendedMethod}}.
}
\details{
This class inherits from the \code{\linkS4class{Component}} class.
}
