% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric-class.R
\name{new_metric}
\alias{new_metric}
\title{Create a Metric object}
\usage{
new_metric(name, label, metric)
}
\arguments{
\item{name}{a short name identifier.  Must be alphanumeric.}

\item{label}{a longer, human readable label that can have other characters
such as spaces, hyphens, etc.}

\item{metric}{a function with arguments "model" and "out" (and optionally
"draw")}
}
\description{
Creates a new \code{\linkS4class{Metric}} object.
}
