% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-from-simulation.R
\name{outputs_or_evals}
\alias{outputs_or_evals}
\title{Internal function used by both outputs and evals}
\usage{
outputs_or_evals(
  sim,
  refs,
  sort_by_method,
  subset,
  index,
  methods,
  reference,
  ...
)
}
\arguments{
\item{sim}{simulation object}

\item{refs}{either sim@output_refs or sim@evals_refs}

\item{sort_by_method}{whether returned object should have each method's objects
in its own list or not}
}
\description{
Internal function used by both outputs and evals
}
\keyword{internal}
