% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{pad}
\alias{pad}
\title{Pad strings in a character vector to the same length with leading zeros}
\usage{
pad(s, w = max(nchar(s)))
}
\arguments{
\item{s}{character vector to pad}

\item{w}{width to pad to}
}
\value{
padded character vector
}
\description{
Pad strings in a character vector to the same length with leading zeros
}
