% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameter_range}
\alias{parameter_range}
\title{Parameter range construction}
\usage{
parameter_range(x, n)
}
\arguments{
\item{x}{points to include in the range, or a list of such vectors}

\item{n}{number of total points filled in by interpolating between x}
}
\value{
vector of parameters of length n
}
\description{
Parameter range will be composed of n points including those
in the vector x.  Sub-lists are processed via recursion using purrr::map
}
