% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameter_range_impl}
\alias{parameter_range_impl}
\title{Insert numbers between vector values until the vector contains
n values}
\usage{
parameter_range_impl(x, n)
}
\arguments{
\item{x}{a vector of numbers}

\item{n}{number of total points}
}
\value{
a vector of length n

determine where the new numbers will occur
determine and insert values of new numbers
}
\description{
find an equal number of equidistant points between each
pair of numbers in x to get a total n number of numbers.
any extra points will be made where the interval
(distance between two #s in x) is largest.
}
