% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{parameters}
\alias{parameters}
\title{Combine environments in a hierarchy for passing simulation data}
\usage{
parameters(
  output = recipe(),
  shared = recipe(),
  initialization = recipe(),
  running = recipe(),
  inputs = list(),
  .parent = rlang::env()
)
}
\arguments{
\item{output}{recipe used for defining output location parameters}

\item{shared}{recipe used for defining shared parameters}

\item{initialization}{recipe used for defining init parameters}

\item{running}{recipe used for defining run-time parameters}

\item{inputs}{list of variables passed in to all steps}

\item{.parent}{dev-side hook, parent of everything.}
}
\value{
list of environments with specified nesting
}
\description{
Combine environments in a hierarchy for passing simulation data
}
