% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening.R
\name{picker_factory}
\alias{picker_factory}
\title{Factory for functions that pick individuals for testing}
\usage{
picker_factory(coverage, interval_days, interval_retest, ...)
}
\arguments{
\item{coverage}{what proportion of rows should be tested}

\item{how}{many days between rounds of testing}

\item{delay}{time required between tests}
}
\value{
function that takes a state df and returns TRUE/FALSE vector for
testing inclusion
}
\description{
Factory for functions that pick individuals for testing
}
