% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eval.R
\name{plot_eval}
\alias{plot_eval}
\title{Make a boxplot of a metric for each method}
\usage{
plot_eval(object, metric_name, use_ggplot2 = TRUE, main, facet_mains, ylab,
  ylim, include_zero = FALSE, angle = 0, ...)
}
\arguments{
\item{object}{an object of class \code{\linkS4class{Simulation}},
\code{\linkS4class{Evals}}, or \code{listofEvals}}

\item{metric_name}{the name of a metric to plot}

\item{use_ggplot2}{whether to use \code{ggplot2} (requires installation
of \code{ggplot2})}

\item{main}{title of plot. Default is \code{model_label} when evals is
a single Evals.}

\item{facet_mains}{only to be used when evals is a \code{listofEvals}
and should be of the same length. Default will be the model_label
for each model.}

\item{ylab}{the y-axis label (default is \code{metric_label})}

\item{ylim}{the y-axis limits to use (across all plots)}

\item{include_zero}{whether ylim should include 0.  Ignored if ylim
is passed explicitly}

\item{angle}{angle of labels (only when \code{use_ggplot2 = FALSE})}

\item{...}{additional arguments to pass to \code{boxplot} (only when
\code{use_ggplot2 = FALSE}).}
}
\description{
If evals is a \code{listofEvals}, then each model will be its own plot.
}

