% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-extension-class.R
\name{+,ExtendedMethod,MethodExtension-method}
\alias{+,ExtendedMethod,MethodExtension-method}
\title{Create an ExtendedMethod from an ExtendedMethod and MethodExtension}
\usage{
\S4method{+}{ExtendedMethod,MethodExtension}(e1, e2)
}
\arguments{
\item{e1}{an object of class \code{\linkS4class{ExtendedMethod}}}

\item{e2}{an object of class \code{\linkS4class{MethodExtension}}}
}
\description{
Create an ExtendedMethod from an ExtendedMethod and MethodExtension
}
