% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-extension-class.R
\name{+,list,MethodExtension-method}
\alias{+,list,MethodExtension-method}
\title{Create a list of ExtendedMethod from a list of Methods and a MethodExtension}
\usage{
\S4method{+}{list,MethodExtension}(e1, e2)
}
\arguments{
\item{e1}{a list of objects of class \code{\linkS4class{Method}} or of class
\code{\linkS4class{ExtendedMethod}}}

\item{e2}{an object of class \code{\linkS4class{MethodExtension}}}
}
\description{
Create a list of ExtendedMethod from a list of Methods and a MethodExtension
}
