% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-extension-class.R
\docType{methods}
\name{+,Method,MethodExtension-method}
\alias{+,Method,MethodExtension-method}
\title{Create an ExtendedMethod from a Method and MethodExtension}
\usage{
\S4method{+}{Method,MethodExtension}(e1, e2)
}
\arguments{
\item{e1}{an object of class \code{\linkS4class{Method}}}

\item{e2}{an object of class \code{\linkS4class{MethodExtension}}}
}
\description{
Create an ExtendedMethod from a Method and MethodExtension
}

