% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{population}
\alias{population}
\title{Container for simulation state, includes states and covariates, supports limited (but
fast) operations required for simulation}
\description{
Supported operations:
}
\details{
Methods that evaluate a dotlist of expressions (matches, mutate, etc...) are
evaluated with a data mask constructed from the data enviornment, and the
listed quosures are individually evaluated in the context of that data mask.
For mutate, this means that expressions that create new data items can be
evaluated sequentially.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{id}}{ids for all members/units}

\item{\code{time}}{current time step, or set current time step}

\item{\code{age}}{population age}

\item{\code{group}}{population group}

\item{\code{states}}{names of population states}

\item{\code{covariates}}{names of population covariates}

\item{\code{n_states}}{number of states in the population}

\item{\code{n_covariates}}{number of covariates in the population}

\item{\code{n_units}}{number of units (rows) in the population}

\item{\code{list}}{return state as a list}

\item{\code{tibble}}{returns state as a tibble}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{population$new()}}
\item \href{#method-get}{\code{population$get()}}
\item \href{#method-set}{\code{population$set()}}
\item \href{#method-absorb}{\code{population$absorb()}}
\item \href{#method-bleb}{\code{population$bleb()}}
\item \href{#method-drop}{\code{population$drop()}}
\item \href{#method-matches}{\code{population$matches()}}
\item \href{#method-mutate}{\code{population$mutate()}}
\item \href{#method-summarize}{\code{population$summarize()}}
\item \href{#method-which}{\code{population$which()}}
\item \href{#method-clone}{\code{population$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{population$new(
  time,
  group,
  data = list(),
  .env = rlang::env(),
  .counter = simulator:::get_counter()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time}}{current time for population}

\item{\code{group}}{population group identifier}

\item{\code{data}}{list of states and covariates, states will be deep-copied to
avoid weird reference semantics in common usage.}

\item{\code{.env}}{environment where the state will be stored (dev-facing
hook)}

\item{\code{.counter}}{object that can generate unique ids (dev-facing hook)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
population object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\if{latex}{\out{\hypertarget{method-get}{}}}
\subsection{Method \code{get()}}{
Get the value of the named population variable for all individuals
(rows)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{population$get(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{data member name to retrieve}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
specified data member
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set"></a>}}
\if{latex}{\out{\hypertarget{method-set}{}}}
\subsection{Method \code{set()}}{
Set the value of the named population variable for all individuals
(rows)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{population$set(name, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{data member name to set}

\item{\code{value}}{value to set the member to}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self, modified
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-absorb"></a>}}
\if{latex}{\out{\hypertarget{method-absorb}{}}}
\subsection{Method \code{absorb()}}{
Merge another population into this one, times must be matched and the
group of the other population will be over-written.  Data must match
between populations (other will be appended).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{population$absorb(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{other populatino to absorb into this one}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self, with other oppended
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bleb"></a>}}
\if{latex}{\out{\hypertarget{method-bleb}{}}}
\subsection{Method \code{bleb()}}{
Separate a set of rows specified by rows or levels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{population$bleb(rows, levels)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rows}}{vector of row indexes to drop}

\item{\code{levels}}{vector of levels (of any state) to drop}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
copy of self, with only specified rows
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-drop"></a>}}
\if{latex}{\out{\hypertarget{method-drop}{}}}
\subsection{Method \code{drop()}}{
Drop specified rows or levels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{population$drop(rows, levels)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rows}}{vector of row indexes to drop}

\item{\code{levels}}{vector of levels (of any state) to drop}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self, without specified rows
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-matches"></a>}}
\if{latex}{\out{\hypertarget{method-matches}{}}}
\subsection{Method \code{matches()}}{
Return boolean vector for rows matching the expression
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{population$matches(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{expressions}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self boolean vector with one value per row
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mutate"></a>}}
\if{latex}{\out{\hypertarget{method-mutate}{}}}
\subsection{Method \code{mutate()}}{
Add covariates or states from a series of evaluated expressions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{population$mutate(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{expressions}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self, for chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summarize"></a>}}
\if{latex}{\out{\hypertarget{method-summarize}{}}}
\subsection{Method \code{summarize()}}{
Evaluate an expression in the context of the population and return the
result
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{population$summarize(..., .return_type = "env")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{expressions}

\item{\code{.return_type}}{how to return evaluated expressions (use a list if
the value is not 'env'}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list or enviornment with evaluated expressions
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-which"></a>}}
\if{latex}{\out{\hypertarget{method-which}{}}}
\subsection{Method \code{which()}}{
Return indexes of rows matching the expression
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{population$which(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{expressions}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
vector of indexes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{population$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
