% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{recipe}
\alias{recipe}
\title{Function for recording a set of program steps}
\usage{
recipe(...)
}
\arguments{
\item{...}{named expressions}
}
\value{
a \code{steps} object
}
\description{
Function for recording a set of program steps
}
