% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eval_by.R
\name{recycle}
\alias{recycle}
\title{Recycles elements to create vector of desired length}
\usage{
recycle(x, length)
}
\arguments{
\item{x}{vector to be expanded to proper length}

\item{length}{desired length}
}
\description{
Recycles elements to create vector of desired length
}
