% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-class.R
\name{relabel}
\alias{relabel}
\title{Give simulation a new label}
\usage{
relabel(sim, label)
}
\arguments{
\item{sim}{object of class \code{\linkS4class{Simulation}}}

\item{label}{a longer, human readable label that can have other characters
such as spaces, hyphens, etc.}
}
\description{
Note that \code{\link{save_simulation}} needs to be called for this change
to be saved to file.
}
\seealso{
\code{\link{rename}}
}

