% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection.R
\name{risk}
\alias{risk}
\title{Risk calculation of SEIR model}
\usage{
risk(state, parameters)
}
\arguments{
\item{state}{as defined by \code{make_susceptible} function}

\item{parameters}{must include 'beta', 'epsilon', and 'gamma'
for infection rate, onset rate, and recovery/death rate}
}
\value{
tibble with p_infection, p_onset, p_recover
}
\description{
Risk calculation of SEIR model
}
