% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{run_extendedmethod_single}
\alias{run_extendedmethod_single}
\title{Run a single extended method on a single index of simulated data.}
\usage{
run_extendedmethod_single(extmethod, model, draws, base_output_list)
}
\arguments{
\item{extmethod}{a \code{\linkS4class{ExtendedMethod}} object}

\item{model}{a \code{\linkS4class{Model}} object}

\item{draws}{a \code{\linkS4class{Draws}} object generated by \code{model}}

\item{base_output_list}{the result of loading a \code{\linkS4class{Output}} object with
\code{more_info = TRUE} so that it includes RNG endstate.}
}
\description{
This is an internal function.  Users should call the wrapper function.
\code{\link{run_method}}. Here "single" refers to a single
index-ExtendedMethod
pair.
}

