% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{run_method_single}
\alias{run_method_single}
\title{Run a single method on a single index of simulated data.}
\usage{
run_method_single(method, model, draws_list)
}
\arguments{
\item{method}{a \code{\linkS4class{Method}} object}

\item{model}{a \code{\linkS4class{Model}} object}

\item{draws_list}{the result of loading a \code{\linkS4class{Draws}} object with
\code{more_info = TRUE} so that it includes RNG endstate.}
}
\description{
This is an internal function.  Users should call the wrapper function.
\code{\link{run_method}}. Here "single" refers to a single index-method
pair.
}

