% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{run_simulation}
\alias{run_simulation}
\title{Run a full set of simulations on the given cluster.}
\usage{
run_simulation(
  parameters,
  initialization,
  step,
  writer,
  exports,
  .env = .GlobalEnv,
  setup = recipe()
)
}
\value{
same parameters as used
}
\description{
Run a full set of simulations on the given cluster.
}
