% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{save_rds}
\alias{save_rds}
\title{Standard saving for simulation vars}
\usage{
save_rds(variable, root, simulation, replicate, time)
}
\arguments{
\item{variable}{variable to save (used for name)}

\item{simulation}{id of simulation, character vector of length 1}

\item{replicate}{id of replicate, character vector of length 1}

\item{time}{id of timepoint, character vector of length 1}

\item{extension}{extension of file type (.rds by default)}

\item{format}{glue package format string}

\item{padding}{how many digits to pad the numeric id's to}
}
\value{
file name
}
\description{
The files are named after the variable saved, with aux id data, they
are saved in .rds format at the given path.
}
