% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-class.R
\name{save_simulation}
\alias{save_simulation}
\title{Save a simulation object}
\usage{
save_simulation(sim)
}
\arguments{
\item{sim}{an object of class \code{\linkS4class{Simulation}}}
}
\description{
Saves an object of class \code{\linkS4class{Simulation}} to
sim@dir/files/sim@name.Rdata. Note: while "files" is the default, the name
of this directory is from getOption("simulator.files"), which is the value of
getOption("simulator.files") when the model was created.
}
\details{
This function overwrites any pre-existing file in that location without
apology.
}
\seealso{
\code{\link{new_simulation}} \code{\link{load_simulation}}
}

