% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws.R
\name{simulate_from_model_single}
\alias{simulate_from_model_single}
\title{Simulate from a model.}
\usage{
simulate_from_model_single(model, nsim, index, seed)
}
\arguments{
\item{model}{a Model object}

\item{nsim}{number of simulations to be conducted.}

\item{index}{a positive integer index.}

\item{seed}{this is the 7 digit seed used by L'Ecuyer RNG}
}
\description{
This is an internal function.  Users should call the wrapper function
\code{\link{simulate_from_model}}.
}

