% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{smooth}
\alias{smooth}
\title{Smooth points at at fine scale}
\usage{
smooth(x, t)
}
\arguments{
\item{x}{values to smooth}

\item{t}{their locations}
}
\value{
smoothed x values
}
\description{
Smooth points at at fine scale
}
