% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{state_files}
\alias{state_files}
\title{Get state-at-time files for a directory}
\usage{
state_files(path)
}
\arguments{
\item{path}{root path to check below}
}
\value{
tibble with directory and a list of files in each directory
}
\description{
They are grouped by replicate directory
}
