% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{steps}
\alias{steps}
\title{Records a set of steps to apply to a simulation later}
\description{
Records a set of steps to apply to a simulation later

Records a set of steps to apply to a simulation later
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{n_steps}}{Number of steps in the recipe}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{steps$new()}}
\item \href{#method-execute}{\code{steps$execute()}}
\item \href{#method-clone}{\code{steps$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Record a list of steps
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{steps$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{expressions to record (argument names are the created targets}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
class storing the expressions
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-execute"></a>}}
\if{latex}{\out{\hypertarget{method-execute}{}}}
\subsection{Method \code{execute()}}{
Execute the stored steps in the pre-specified environment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{steps$execute(bottom = rlang::env(), top = bottom, target = bottom)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bottom}}{other end of environment stack to execute in}

\item{\code{top}}{one end of environment stack to execute in}

\item{\code{target}}{environment to place recipe results in (if named).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
modified environment
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{steps$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
