% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{subset_evals}
\alias{subset_evals}
\title{Reduce an Evals object to a subset of methods and/or metrics}
\usage{
subset_evals(evals, method_names = NULL, metric_names = NULL)
}
\arguments{
\item{evals}{an object of class \code{\linkS4class{Evals}} or
\code{listofEvals}.}

\item{method_names}{a character vector of method names}

\item{metric_names}{a character vector of metric names}
}
\description{
If \code{method_names} is \code{NULL}, then subsetting is not done over
methods.  Likewise for \code{metric_names}.
}
