% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\name{subset_models}
\alias{subset_models}
\title{Subset Models}
\usage{
subset_models(m, ...)
}
\arguments{
\item{m}{list of \code{\linkS4class{Model}} objects}

\item{...}{logical expression involving parameters of Models.  For now, can
only be parameters that are of length 1 and either of class numeric or
character}
}
\description{
Given a list of \code{\linkS4class{Model}} objects, returns model names
which meet conditions.  Uses \code{\link{subset}}
}
