% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{summarize_simulation_state}
\alias{summarize_simulation_state}
\title{Summarize states for whole simulation directory}
\usage{
summarize_simulation_state(path, env = .GlobalEnv)
}
\arguments{
\item{path}{process the entire tree rooted at the path}

\item{env}{additional data loaded on each node}
}
\value{
list of state summary files generated
}
\description{
Each simulation is in one directory
}
