% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{summarize_state_file}
\alias{summarize_state_file}
\title{Summarize state file}
\usage{
summarize_state_file(file)
}
\arguments{
\item{file}{file to summarize}
}
\value{
specified summary
}
\description{
Load file and defer to fixed summary function(s)
}
