% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timings.R
\name{summarize_timings}
\alias{summarize_timings}
\title{Summarize timings}
\usage{
summarize_timings(timings)
}
\arguments{
\item{timing}{data loaded with \code{get_timings()}}
}
\value{
tibble of timings with mean/sd/etc..
}
\description{
Summarizes across all replicates
}
