% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{summary_smooth}
\alias{summary_smooth}
\title{Smooth states over time within each location}
\usage{
summary_smooth(data)
}
\arguments{
\item{data}{state counts}
}
\value{
smoothed states by set and location
}
\description{
Smooth states over time within each location
}
