% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{tag_picks}
\alias{tag_picks}
\title{Of n items, only those at idx \code{picks} are TRUE}
\usage{
tag_picks(picks, n)
}
\arguments{
\item{picks}{these will be TRUE, the rest FALSE}

\item{n}{total number of items}
}
\value{
a vector of logical values
}
\description{
Of n items, only those at idx \code{picks} are TRUE
}
