% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening.R
\name{test_factory}
\alias{test_factory}
\title{Generic test factory}
\usage{
test_factory(sensitivity, specificity, ...)
}
\arguments{
\item{sensitivity, }{as per usual}

\item{specificity, }{as per usual}
}
\value{
a test function with given specificity/sensitivity
}
\description{
Generic test factory
}
