% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{two_state_generator}
\alias{two_state_generator}
\title{Generate a mix of two states, based on probability of the first state}
\usage{
two_state_generator(n, p, labels, .state = TRUE)
}
\arguments{
\item{n}{number of units to generate}

\item{p}{probability of first label}

\item{labels}{two labels for the two stats}
}
\value{
n-vector of either of two labels
}
\description{
Generate a mix of two states, based on probability of the first state
}
