% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{writer}
\alias{writer}
\title{Write that records simulation state}
\description{
Write that records simulation state

Write that records simulation state
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{writer$new()}}
\item \href{#method-setup}{\code{writer$setup()}}
\item \href{#method-record}{\code{writer$record()}}
\item \href{#method-save}{\code{writer$save()}}
\item \href{#method-finalize}{\code{writer$finalize()}}
\item \href{#method-clone}{\code{writer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a writer...
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{writer$new(setup = recipe(), summaries = recipe())}\if{html}{\out{</div>}}
}

\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setup"></a>}}
\if{latex}{\out{\hypertarget{method-setup}{}}}
\subsection{Method \code{setup()}}{
Set up the writer for a specific task, the enviroment must provide
all the specific parameters used
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{writer$setup(bottom, top)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-record"></a>}}
\if{latex}{\out{\hypertarget{method-record}{}}}
\subsection{Method \code{record()}}{
Record data from the environment after applying summaries
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{writer$record(bottom, top)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save"></a>}}
\if{latex}{\out{\hypertarget{method-save}{}}}
\subsection{Method \code{save()}}{
Save recorded data...
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{writer$save()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{writer$finalize(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{writer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
