% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{generate_jobs}
\alias{generate_jobs}
\title{Turn inputs into a set of job descriptors}
\usage{
generate_jobs(
  env,
  modifiers,
  n_replicates,
  output_root = "output",
  tests = test_factory,
  pickers = picker_factory
)
}
\arguments{
\item{env}{an environment where constant parameters are defined}

\item{modifiers}{a list of tibbles where per-scenario parameters are defined}

\item{n_replicates}{number of replicates to create per set}

\item{output_root}{location to output results to.  There is a specific
directory structure implied under this location (it will be created).}

\item{tests}{a factory function for creating testing functions for each set
by default uses functions from the package}

\item{pickers}{a factory function for createing picker functions for each
set, by default uses factory function from the package}
}
\value{
a list of list with a descriptor of sets, sets with replicates, and
a list of jobs to run
}
\description{
Turn inputs into a set of job descriptors
}
