% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{get_contents}
\alias{get_contents}
\title{Get the contents of a simulator directory}
\usage{
get_contents(dir = ".", out_loc = "out")
}
\arguments{
\item{dir}{name of the directory where directory named "files" exists}

\item{out_loc}{a length-1 character vector that gives location
(relative to model's path) that method outputs are stored.This can be
useful for staying organized when multiple simulations are based on
the same Model and Draws objects.  Usually this is just "out"}
}
\description{
Get the contents of a simulator directory
}

