% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameters}
\alias{parameters}
\title{Update parameter data with stage-specific recipes}
\usage{
parameters(
  data = list(),
  output = recipe(),
  shared = recipe(),
  initialization = recipe(),
  running = recipe(),
  .parent = rlang::env(),
  .return_type = "list"
)
}
\arguments{
\item{output}{recipe used for defining output location parameters}

\item{shared}{recipe used for defining shared parameters}

\item{initialization}{recipe used for defining init parameters}

\item{running}{recipe used for defining run-time parameters}

\item{.parent}{dev-side hook, parent of everything.}

\item{.data}{list of variables passed in to all steps}
}
\value{
list of environments with specified nesting
}
\description{
Input data is present in the root environment and further
stages form a hierarchy for the derived stage-specific data
defined in recipes
}
