% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{simulation}
\alias{simulation}
\title{Run a simulation for a single set of parameters}
\usage{
simulation(parameters, initialization, step, writer = simulator::writer$new())
}
\arguments{
\item{parameters}{see \code{parameters} function}

\item{initialization}{recipe for initializing sim population}

\item{step}{recipe for runnning the simulation}

\item{writer}{for recording simulation summaries}
}
\value{
writer, post-application
}
\description{
Run a simulation for a single set of parameters
}
